use subcommand defaults instead of action map
The action_map was a workaround for the problems caused by the fact that values set via set_defaults for a subcommand override those for any nested subcommands.
Since it's not possible to use optional subcommands, they are not suitable for the parser, so I'm getting rid of the action map.